+++ pya/Asig.py
@@ -1,4 +1,4 @@
-from .Aserver import Aserver
+
 import numbers
 from warnings import warn
 import logging
@@ -9,7 +9,11 @@ import scipy.interpolate
 import scipy.signal
 from scipy.fftpack import fft, fftfreq, ifft
 from scipy.io import wavfile
-from .helpers import ampdb, dbamp, linlin, timeit, spectrum, audio_from_file
+from . import Aserver
+from . import Aspec
+from . import Astft
+from .helpers import ampdb, dbamp, cpsmidi, midicps, linlin, clip, record, timeit, audio_from_file, buf_to_float
+from .helpers import spectrum, audio_from_file
 from copy import copy, deepcopy
 
 _LOGGER = logging.getLogger(__name__)
@@ -542,9 +546,9 @@ class Asig:
             s = kwargs['server']
         else:
             s = Aserver.default
-        if not isinstance(s, Aserver):
+        if not isinstance(s, Aserver.Aserver):
             warn("Asig.play: no default server running, nor server arg specified.")
-            return
+            return self
         if rate == 1 and self.sr == s.sr:
             asig = self
         else:
@@ -1098,6 +1102,7 @@ class Asig:
             and self._['a']
 
         """
+        # TODO scipy.signal.__getattribute__ error
         Wn = np.array(cutoff_freqs) * 2 / self.sr
         b, a = scipy.signal.iirfilter(order, Wn, rp=rp, rs=rs, btype=btype, ftype=ftype)
         y = scipy.signal.__getattribute__(filter)(b, a, self.sig, axis=0)
@@ -1467,275 +1472,4 @@ class Asig:
         """custom function method. TODO add example"""
         func(self, **kwargs)
         return self
-
-
-class Aspec:
-    """Audio spectrum class using rfft"""
-    def __init__(self, x, sr=44100, label=None, cn=None):
-        """__init__() method
-        Parameters
-        ----------
-        x : Asig or numpy.ndarray
-            audio signal
-        sr : int
-            sampling rate (Default value = 44100)
-        label : str or None
-            Asig label (Default value = None)
-        cn : list or None
-            Channel names (Default value = None)
-        """
-        self.cn = cn
-        if type(x) == Asig:
-            self.sr = x.sr
-            self.rfftspec = np.fft.rfft(x.sig)
-            self.label = x.label + "_spec"
-            self.samples = x.samples
-            self.channels = x.channels
-            self.cn = x.cn
-            if cn is not None and self.cn != cn:
-                warn("Aspec:init: given cn  different from Asig cn: using Asig.cn")
-        elif type(x) == list or type(x) == np.ndarray:
-            self.rfftspec = np.array(x)
-            self.sr = sr
-            self.samples = (len(x) - 1) * 2
-            self.channels = 1
-            if len(np.shape(x)) > 1:
-                self.channels = np.shape(x)[1]
-        else:
-            raise AttributeError("unknown initializer")
-        if label:
-            self.label = label
-        self.nr_freqs = self.samples // 2 + 1
-        self.freqs = np.linspace(0, self.sr / 2, self.nr_freqs)
-
-    def to_sig(self):
-        """Convert Aspec into Asig"""
-        return Asig(np.fft.irfft(self.rfftspec), sr=self.sr, label=self.label + '_2sig', cn=self.cn)
-
-    def weight(self, weights, freqs=None, curve=1, kind='linear'):
-        """TODO
-
-        Parameters
-        ----------
-        weights :
-
-        freqs :
-             (Default value = None)
-        curve :
-             (Default value = 1)
-        kind :
-             (Default value = 'linear')
-
-        Returns
-        -------
-
-        """
-        nfreqs = len(weights)
-        if not freqs:
-            given_freqs = np.linspace(0, self.freqs[-1], nfreqs)
-        else:
-            if nfreqs != len(freqs):
-                _LOGGER.error("len(weights)!=len(freqs)")
-                return self
-            if all(freqs[i] < freqs[i + 1] for i in range(len(freqs) - 1)):  # check if list is monotonous
-                if freqs[0] > 0:
-                    freqs = np.insert(np.array(freqs), 0, 0)
-                    weights = np.insert(np.array(weights), 0, weights[0])
-                if freqs[-1] < self.sr / 2:
-                    freqs = np.insert(np.array(freqs), -1, self.sr / 2)
-                    weights = np.insert(np.array(weights), -1, weights[-1])
-            else:
-                _LOGGER.error("Aspec.weight error: freqs not sorted")
-                return self
-            given_freqs = freqs
-        if nfreqs != self.nr_freqs:
-            interp_fn = scipy.interpolate.interp1d(given_freqs, weights, kind=kind)
-            rfft_new = self.rfftspec * interp_fn(self.freqs) ** curve  # ToDo: curve segmentwise!!!
-        else:
-            rfft_new = self.rfftspec * weights ** curve
-        return Aspec(rfft_new, self.sr, label=self.label + "_weighted")
-
-    def plot(self, fn=np.abs, xlim=None, ylim=None, **kwargs):  # TODO add ax option
-        """Plot spectrum
-
-        Parameters
-        ----------
-        fn : func
-            function for processing the rfft spectrum. (Default value = np.abs)
-        xlim : tuple or list or None
-            Set x axis range (Default value = None)
-        ylim : tuple or list or None
-            Set y axis range (Default value = None)
-        **kwargs :
-            Keyword arguments for matplotlib.pyplot.plot()
-
-        Returns
-        -------
-        _ : Asig
-            self
-        """
-        plt.plot(self.freqs, fn(self.rfftspec), **kwargs)
-        if xlim is not None:
-            plt.xlim([xlim[0], xlim[1]])
-
-        if ylim is not None:
-            plt.ylim([ylim[0], ylim[1]])
-
-        plt.xlabel('freq (Hz)')
-        plt.ylabel(f'{fn.__name__}(freq)')
-        return self
-
-    def __repr__(self):
-        return "Aspec('{}'): {} x {} @ {} Hz = {:.3f} s".format(
-            self.label, self.channels, self.samples, self.sr, self.samples / self.sr)
-
-
-# TODO, check with multichannel
-class Astft:
-    """Audio spectrogram (STFT) class, attributes refers to scipy.signal.stft. With an addition
-        attribute cn being the list of channel names, and label being the name of the Asig
-    """
-
-    def __init__(self, x, sr=None, label=None, window='hann', nperseg=256,
-                 noverlap=None, nfft=None, detrend=False, return_onesided=True,
-                 boundary='zeros', padded=True, axis=-1, cn=None):
-        """__init__() method
-
-        Parameters
-        ----------
-        x : Asig or numpy.ndarray
-            signal to be converted to stft domain. This can be either a numpy array or an Asig object. 
-        sr : int
-            sampling rate, this is only necessary if x is not Asig. (Default value = None)
-        label : str
-            name of the Asig. (Default value = None)
-        window : str
-            type of the window function (Default value = 'hann')
-        nperseg : int
-            number of samples per stft segment (Default value = '256')
-        noverlap : int
-            number of samples to overlap between segments (Default value = None)
-        detrend : str or function or bool
-            Specifies how to detrend each segment. If detrend is a string, 
-            it is passed as the type argument to the detrend function. If it is a function, 
-            it takes a segment and returns a detrended segment. If detrend is False, 
-            no detrending is done. (Default value = False).
-        return_onesided : bool
-            If True, return a one-sided spectrum for real data. If False return a two-sided spectrum. 
-            Defaults to True, but for complex data, a two-sided spectrum is always returned. (Default value = True)
-        boundary : str or None
-            Specifies whether the input signal is extended at both ends, and how to generate the new values, 
-            in order to center the first windowed segment on the first input point. 
-            This has the benefit of enabling reconstruction of the first input point 
-            when the employed window function starts at zero. 
-            Valid options are ['even', 'odd', 'constant', 'zeros', None]. Defaults to ‘zeros’, 
-            for zero padding extension. I.e. [1, 2, 3, 4] is extended to [0, 1, 2, 3, 4, 0] for nperseg=3. (Default value = 'zeros')
-        padded : bool
-            Specifies whether the input signal is zero-padded at the end to make the signal fit exactly into 
-            an integer number of window segments, so that all of the signal is included in the output. 
-            Defaults to True. Padding occurs after boundary extension, if boundary is not None, and padded is True, 
-            as is the default. (Default value = True)
-        axis : int
-            Axis along which the STFT is computed; the default is over the last axis. (Default value = -1)
-        cn : list or None
-            Channel names of the Asig, this will be used for the Astft for consistency. (Default value = None)
-        """
-        self.window = window
-        self.nperseg = nperseg
-        self.noverlap = noverlap
-        self.nfft = nfft
-        self.detrend = detrend
-        self.return_onesided = return_onesided
-        self.boundary = boundary
-        self.padded = padded
-        self.axis = axis
-        self.cn = cn
-        if type(x) == Asig:
-            # TODO multichannel.
-            self.sr = x.sr
-            if sr:
-                self.sr = sr  # explicitly given sr overwrites Asig
-            self.freqs, self.times, self.stft = scipy.signal.stft(
-                x.sig, fs=self.sr, window=window, nperseg=nperseg, noverlap=noverlap, nfft=nfft,
-                detrend=detrend, return_onesided=return_onesided, boundary=boundary, padded=padded, axis=axis)
-            self.label = x.label + "_stft"
-            self.samples = x.samples
-            self.channels = x.channels
-        elif type(x) == np.ndarray and np.shape(x) >= 2:
-            self.stft = x
-            self.sr = 44100
-            if sr:
-                self.sr = sr
-            self.samples = (len(x) - 1) * 2
-            self.channels = 1
-            if len(np.shape(x)) > 2:
-                self.channels = np.shape(x)[2]
-            # TODO: set other values, particularly check if self.times and self.freqs are correct
-            self.ntimes, self.nfreqs, = np.shape(self.stft)
-            self.times = np.linspace(0, (self.nperseg - self.noverlap) * self.ntimes / self.sr, self.ntimes)
-            self.freqs = np.linspace(0, self.sr // 2, self.nfreqs)
-        else:
-            print("error: unknown initializer or wrong stft shape ")
-        if label:
-            self.label = label
-
-    def to_sig(self, **kwargs):
-        """Create signal from stft, i.e. perform istft, kwargs overwrite Astft values for istft
-
-        Parameters
-        ----------
-        **kwargs : str
-            optional keyboard arguments used in istft: 
-                'sr', 'window', 'nperseg', 'noverlap', 'nfft', 'input_onesided', 'boundary'.
-            also convert 'sr' to 'fs' since scipy uses 'fs' as sampling frequency.
-
-        Returns
-        -------
-        _ : Asig
-            Asig
-        """
-        for k in ['sr', 'window', 'nperseg', 'noverlap', 'nfft', 'input_onesided', 'boundary']:
-            if k in kwargs.keys():
-                kwargs[k] = self.__getattribute__(k)
-        if 'sr' in kwargs.keys():
-            kwargs['fs'] = kwargs['sr']
-            del kwargs['sr']
-        _, sig = scipy.signal.istft(self.stft, **kwargs)  # _ since 1st return value 'times' unused
-        return Asig(sig, sr=self.sr, label=self.label + '_2sig', cn=self.cn)
-
-    def plot(self, fn=lambda x: x, ax=None, xlim=None, ylim=None, **kwargs):
-        """Plot spectrogram
-
-        Parameters
-        ----------
-        fn : func
-            a function, by default is bypass
-        ax : matplotlib.axes
-            you can assign your plot to specific axes (Default value = None)
-        xlim : tuple or list
-            x_axis range (Default value = None)
-        ylim : tuple or list
-            y_axis range (Default value = None)
-        **kwargs :
-            keyward arguments of matplotlib's pcolormesh
-
-        Returns
-        -------
-        _ : Asig
-            self
-        """
-        if ax is None:
-            plt.pcolormesh(self.times, self.freqs, fn(np.abs(self.stft)), **kwargs)
-            plt.colorbar()
-            if ylim is not None:
-                plt.ylim([ylim[0], ylim[1]])
-        else:
-            ax.pcolormesh(self.times, self.freqs, fn(np.abs(self.stft)), **kwargs)
-            if ylim is not None:
-                ax.set_ylim(ylim[0], ylim[1])
-        return self
-
-    def __repr__(self):
-        return "Astft('{}'): {} x {} @ {} Hz = {:.3f} s".format(
-            self.label, self.channels, self.samples, self.sr, self.samples / self.sr, cn=self.cn)