+++ pya/__init__.py
@@ -1,11 +1,15 @@
 """Collection of classes and functions for processing audio signals
  in python and jupyter notebooks, for synthesis, effects, analysis and plotting.
 """
-from .pya import Asig, Aspec, Astft, Arecorder
+from .Asig import Asig
+from .Astft import Astft
+from .Aspec import Aspec
 from .Aserver import Aserver
+from .Arecorder import Arecorder
 from .Ugen import Ugen
-from .helpers import ampdb, dbamp, cpsmidi, midicps, linlin, clip, record, timeit, audio_from_file, buf_to_float
 from .version import __version__
+
+__all__ = ['Asig', 'Aspec', 'Astft', 'Aserver', 'Arecorder', 'Ugen']
 
 
 def startup(**kwargs):